# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garuka API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ussd/{secret}": {
      "post": {
        "tags": [
          "ussd"
        ],
        "summary": "Africa's Talking USSD Webhook",
        "description": "USSD Webhook handler for Africa's Talking.\nAlways returns HTTP 200 with text/plain, starting with CON or END.\nNever logs raw PINs or raw cumulative text.",
        "operationId": "handle_ussd_callback",
        "parameters": [
          {
            "name": "secret",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_handle_ussd_callback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dev/ussd": {
      "post": {
        "tags": [
          "ussd"
        ],
        "summary": "Dev USSD Simulator endpoint",
        "description": "Dev-only USSD endpoint simulating Africa's Talking webhook.",
        "operationId": "dev_ussd_simulator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevUssdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_handle_ussd_callback": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid",
            "default": ""
          },
          "serviceCode": {
            "type": "string",
            "title": "Servicecode",
            "default": ""
          },
          "phoneNumber": {
            "type": "string",
            "title": "Phonenumber",
            "default": ""
          },
          "networkCode": {
            "type": "string",
            "title": "Networkcode",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          }
        },
        "type": "object",
        "title": "Body_handle_ussd_callback"
      },
...
      "DevUssdRequest": {
        "properties": {
          "phoneNumber": {
            "type": "string",
            "title": "Phonenumber"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "sessionId": {
            "type": "string",
            "title": "Sessionid",
            "default": "dev-session-1"
          },
          "serviceCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servicecode",
            "default": "*384*1234#"
          }
        },
        "type": "object",
        "required": [
          "phoneNumber"
        ],
        "title": "DevUssdRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}